tool-search-tool-quickstart/pyproject.toml
[project]
name = "tool-search-tool-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "ipython>=9.8.0",
  "langchain>=1.1.0",
  "langchain-google-genai>=3.2.0",
  "langchain-mcp-adapters>=0.1.14",
  "mcp>=1.22.0",
  "python-dotenv>=1.2.1",
  "rank-bm25>=0.2.2",
]
